Switch to WSL
