+++ .github/workflows/ci.yml
@@ -24,10 +24,8 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v7.0.1
-      - name: Azure login
-        uses: azure/login@v3.0.1
-        with:
-          creds: ${{ secrets.AZURE_ACI_CREDENTIALS }}
+      - name: Setup WSL
+        uses: Particular/setup-wsl-action@v1.0.0
       - name: Run
         uses: ./
         with:
