+++ action.yml
@@ -26,6 +26,6 @@ inputs:
     required: false
 
 runs:
-  using: node20
-  main: dist/index.js
-  post: dist/index.js
+  using: node24
+  main: dist/index.mjs
+  post: dist/index.mjs
