+++ cleanup.ps1
@@ -2,8 +2,12 @@ param (
     [string]$RabbitMQName
 )
 
-$resourceGroup = $Env:RESOURCE_GROUP_OVERRIDE ?? "GitHubActions-RG"
 $runnerOs = $Env:RUNNER_OS ?? "Linux"
+
+if (-not $Env:WSL_TOOLS_MODULE_PATH) {
+    throw "This action requires Particular/setup-wsl-action to run first — it provisions WSL/Docker and exports the WslTools module at WSL_TOOLS_MODULE_PATH."
+}
+Import-Module $Env:WSL_TOOLS_MODULE_PATH -Force
 
 if ($runnerOs -eq "Linux") {
     Write-Output "Killing Docker container $RabbitMQName"
@@ -13,8 +17,8 @@ if ($runnerOs -eq "Linux") {
     docker rm $RabbitMQName
 }
 elseif ($runnerOs -eq "Windows") {
-    Write-Output "Deleting Azure container $RabbitMQName"
-    az container delete --resource-group $resourceGroup --name $RabbitMQName --yes | Out-Null
+    Write-Output "Removing WSL Docker container $RabbitMQName"
+    Invoke-Wsl -Command "docker rm --force ${RabbitMQName} 2>/dev/null || true"
 }
 else {
     Write-Output "$runnerOs not supported"
